Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-img-element/image-loading-lazy-crossorigin-change.sub-expected.txt

CONSOLE MESSAGE: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Cannot load image http://127.0.0.1:8800/html/semantics/embedded-content/the-img-element/resources/image.png due to access control checks.


PASS Test that when deferred image is loaded, it uses the latest crossorigin attribute.
